Update prog1/week 2/module 5/population.c

+++ prog1/week 2/module 5/population.c
@@ -3,7 +3,7 @@
 
 int main (void)
 {
-    float n;
+    int n;
     int x;
     int years;
